split powercontrol algo
